requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
assemblyai==0.33.0
asttokens==2.4.1
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
executing==2.1.0
Flask==3.0.3
flask-sock==0.7.0
Flask-SocketIO==5.3.7
Flask-Sockets==0.2.1
frozenlist==1.4.1
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.1.0
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
ngrok==1.4.0
numpy==2.1.1
openai==1.47.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
pyzmq==26.2.0
requests==2.32.3
setuptools==75.1.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
twilio==9.3.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
websockets==13.0.1
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.11.1
zope.event==5.0
zope.interface==7.0.3
